Flask APPI
I struggled a lot to do this assignment. I spent all day working on it, and I was finally able to make the app work. I had installed two versions of Python, so that caused me a lot of problems and confusion, so I had to uninstall and reinstall it again only after reading a lot about where to install it or not to have full privileges on all files. I realized that Flask only worked when I created my app and venv through the C:Drive; otherwise, it looked like it didn't work directly from my username. I also had difficulty creating and activating the venv as well as putting all the pieces together within my APP, templates, and static files. I moved things around until it worked just now. I really put the effor doing this assignment because I want to master it.